Resolved #43
