+++ lib/util.js
@@ -228,6 +228,24 @@ function parseVectorObjectToString(obj) {
   return vectorString;
 }
 
+/**
+ * Retrives the version from the vector string
+ * 
+ * @return {String} returns the version number 
+ */
+function getVersion(vector){
+  const version = vector.split("/");
+  if (version[0] === "CVSS:3.0") {
+    return "3.0";
+  }
+  else if (version[0] === "CVSS:3.1") {
+    return "3.1";
+  }
+  else {
+    return "Error";
+  }
+}
+
 module.exports = {
   roundUpExact,
   roundUpApprox,
@@ -237,5 +255,6 @@ module.exports = {
   findMetricValue,
   getRating,
   isVectorValid,
-  parseVectorObjectToString
+  parseVectorObjectToString,
+  getVersion
 };
